[Task 3] feat: add file attachment and email sending
Add the ability to attach a PDF file to emails.
Update `MainConst` to include `WM_ATTACHMENT_URI` for the attachment URI.
Update `SendEmailFragment` to allow the selection of a PDF file and include it as an attachment.
Update `EmailSenderWorker` to handle email sending with file attachments.
Update `EmailViewModel` to support sending emails with attachments.
Modify `fragment_fill_form.xml` to change designation input type to `text`.
Add the functionality to update button state based on file selection.
Add function to check if the fields are empty.
Add file picker functionality.
Add a validation check to ensure a resume is selected before sending mail.
Remove unnecessary comments and replace them with TODO comments.